+++ tests/mpftests/reference_images/addplot01.png
