+++ tests/mpftests/reference_images/addplot02.png
