+++ tox.ini
@@ -6,8 +6,8 @@
 [tox]
 envlist = py36, py37, py38
 
-[pytest]
-python_files = tests.py
+#[pytest]
+#python_files = tests.py
 
 [testenv]
 deps =
